+++ src/main/java/com/challenge/devchall/challange/repository/ChallengeRepository.java
@@ -10,13 +10,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 public interface ChallengeRepository extends JpaRepository<Challenge, Long> {
-        @Query("SELECT cm.id as challengemember_id, COUNT(p.id) as count " +
-            "FROM ChallengeMember cm " +
-            "JOIN cm.linkedChallenge c " +
-            "JOIN ChallengePost p ON cm.challenger = p.challenger AND p.linkedChallenge = c " +
-            "WHERE c.endDate = :today " +
-            "GROUP BY cm.id") // 챌린지 아이디 + 멤버 아이디
-        List<SettleChallengeDTO> findChallengeMemberCountByEndDate(@Param("today") LocalDate today);
+
 
     List<Challenge> findByChallengeStatus(boolean challengeStatus, Pageable pageable);
