+++ src/main/java/com/challenge/devchall/challengeMember/repository/ChallengeMemberRepository.java
@@ -1,10 +1,14 @@
 package com.challenge.devchall.challengeMember.repository;
 
+import com.challenge.devchall.challange.dto.SettleChallengeDTO;
 import com.challenge.devchall.challange.entity.Challenge;
 import com.challenge.devchall.challengeMember.entity.ChallengeMember;
 import com.challenge.devchall.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,5 +19,11 @@ public interface ChallengeMemberRepository extends JpaRepository<ChallengeMember
     List<ChallengeMember> findByLinkedChallenge(Challenge challenge);
 
     Optional<ChallengeMember> findByLinkedChallengeAndChallenger(Challenge challenge, Member member);
-
+    @Query("SELECT cm.id as challengemember_id, COUNT(p.id) as count " +
+            "FROM ChallengeMember cm " +
+            "JOIN cm.linkedChallenge c " +
+            "JOIN ChallengePost p ON cm.challenger = p.challenger AND p.linkedChallenge = c " +
+            "WHERE c.endDate = :today " +
+            "GROUP BY cm.id") // 챌린지 아이디 + 멤버 아이디
+    List<SettleChallengeDTO> findChallengeMemberCountByEndDate(@Param("today") LocalDate today);
 }
